+++ nari/types/event/act/debug.py
