+++ nari/types/event/act/hook.py
