+++ nari/types/event/act/networkeffectresult.py
