+++ nari/types/event/act/networkstatuseffect.py
