+++ nari/types/event/act/removecombatant.py
